fix(vm): lossless closure capture-count encoding (versioned ESKB); prelude cache regenerated; 255/256/257/4096/65536-capture regressions on VM/native/WASM (review P0/P1)

+++ lib/core/arena_memory.h
@@ -239,7 +239,7 @@ void* arena_allocate_symbol_with_header(arena_t* arena, size_t length);
 // Returns pointer to closure data (header is at offset -8)
 // name: bound procedure name from (define name ...) or NULL for anonymous lambdas
 eshkol_closure_t* arena_allocate_closure_with_header(arena_t* arena, uint64_t func_ptr,
-                                                      size_t num_captures, uint64_t sexpr_ptr,
+                                                      uint64_t num_captures, uint64_t sexpr_ptr,
                                                       uint64_t return_type_info,
                                                       const char* name);
 
@@ -782,7 +782,7 @@ eshkol_closure_env_t* arena_allocate_closure_env(arena_t* arena, size_t num_capt
 // Allocate full closure structure (func_ptr + environment + sexpr for homoiconicity)
 // return_type_info: packed return type metadata (return_type | (input_arity << 8) | (hott_type_id << 16))
 // name: bound procedure name from (define name ...) or NULL for anonymous lambdas
-eshkol_closure_t* arena_allocate_closure(arena_t* arena, uint64_t func_ptr, size_t num_captures,
+eshkol_closure_t* arena_allocate_closure(arena_t* arena, uint64_t func_ptr, uint64_t num_captures,
                                          uint64_t sexpr_ptr, uint64_t return_type_info,
                                          const char* name);
